+++ scala-package/examples/src/main/scala/org/apache/mxnetexamples/cnntextclassification/README.md
